+++ Changelog.md
@@ -1,2 +1,6 @@
+## 0.0.2
++ Added `fxn.predictions.ready` method for checking if a predictor is preloaded and ready for making predictions.
++ Added `Bootstrap Project` Xcode project command to add Function configuration files into a new project.
+
 ## 0.0.1
 + First pre-release.